修改为单精度浮点数

+++ example/couplet/logic/sample/build.go
@@ -1,13 +1,13 @@
 package sample
 
-func onehot(x, size int) []float64 {
-	ret := make([]float64, size)
+func onehot(x, size int) []float32 {
+	ret := make([]float32, size)
 	ret[x] = 1
 	return ret
 }
 
-func zerohot(size int) []float64 {
-	ret := make([]float64, size)
+func zerohot(size int) []float32 {
+	ret := make([]float32, size)
 	return ret
 }
 
@@ -24,10 +24,10 @@ func encode(vocabs []string, idx []int) string {
 }
 
 // Build 生成一个样本，输出: sequence, next word, padding mask
-func Build(x []int, y, paddingSize int, embedding [][]float64, vocabs []string) ([]float64, []float64, []bool) {
+func Build(x []int, y, paddingSize int, embedding [][]float32, vocabs []string) ([]float32, []float32, []bool) {
 	// fmt.Printf("%s => %s\n", encode(vocabs, x), encode(vocabs, []int{y}))
 	embeddingSize := len(embedding[0])
-	dx := make([]float64, 0, paddingSize*embeddingSize)
+	dx := make([]float32, 0, paddingSize*embeddingSize)
 	paddingMask := make([]bool, 0, paddingSize)
 	for _, v := range x {
 		dx = append(dx, embedding[v]...)
@@ -39,7 +39,7 @@ func Build(x []int, y, paddingSize int, embedding [][]float64, vocabs []string) 
 		}
 		paddingMask = append(paddingMask, true)
 	}
-	dz := make([]float64, 0, len(embedding))
+	dz := make([]float32, 0, len(embedding))
 	if y < 0 {
 		dz = append(dz, zerohot(len(embedding))...)
 	} else {
