+++ nn/initializer/number.go
@@ -1,27 +1,27 @@
 package initializer
 
 type Number struct {
-	n float64
+	n float32
 }
 
-func NewNumber(n float64) *Number {
+func NewNumber(n float32) *Number {
 	return &Number{n: n}
 }
 
-func (rand *Number) Rand() float64 {
+func (rand *Number) Rand() float32 {
 	return rand.n
 }
 
-func (rand *Number) RandN(n int) []float64 {
-	ret := make([]float64, n)
+func (rand *Number) RandN(n int) []float32 {
+	ret := make([]float32, n)
 	for i := 0; i < n; i++ {
 		ret[i] = 1
 	}
 	return ret
 }
 
-func (rand *Number) RandShape(m, n int) []float64 {
-	ret := make([]float64, m*n)
+func (rand *Number) RandShape(m, n int) []float32 {
+	ret := make([]float32, m*n)
 	for i := 0; i < m*n; i++ {
 		ret[i] = 1
 	}
