+++ nn/initializer/zero.go
@@ -7,14 +7,14 @@ func NewZero() *Zero {
 	return &Zero{}
 }
 
-func (rand *Zero) Rand() float64 {
+func (rand *Zero) Rand() float32 {
 	return 0
 }
 
-func (rand *Zero) RandN(n int) []float64 {
-	return make([]float64, n)
+func (rand *Zero) RandN(n int) []float32 {
+	return make([]float32, n)
 }
 
-func (rand *Zero) RandShape(m, n int) []float64 {
-	return make([]float64, m*n)
+func (rand *Zero) RandShape(m, n int) []float32 {
+	return make([]float32, m*n)
 }
